requirements.txt
aicsimageio==4.14.0
joblib==1.3.2
lightning==2.1.3 
matplotlib==3.8.2
matplotlib-scalebar==0.8.1
n2v==0.3.3
numpy==1.26.3 
pandas==2.2.0
pillow==10.2.0
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
torch==2.1.2 
torch-tb-profiler==0.4.3
torchdata==0.7.1
torchmetrics==1.3.0.post0
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
zarr==2.15.0
quilt3==5.4.0
pyarrow==15.0.0
tensorflow==2.15.0